Codebase improvement, separeted updates from store
