+++ awt-admin/css/pageEditor.css
@@ -6,9 +6,6 @@ body {
     grid-template-columns: auto auto !important;
 }
 
-
-
-
 .page {
     padding: 0;
     margin: 0;
